requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
streamlit>=1.40.0
plotly>=5.18.0
python-docx>=1.0.0
reportlab>=4.0.0
weasyprint>=62.0
psycopg2-binary>=2.9.0
tavily-python>=0.5.0
firecrawl-py>=1.0.0
ollama>=0.1.0
requests>=2.28.0
